+++ cutadapt/report.py
@@ -10,7 +10,8 @@ import textwrap
 from .adapters import BACK, FRONT, PREFIX, SUFFIX, ANYWHERE, LINKED
 from .modifiers import QualityTrimmer, AdapterCutter
 from .filters import (NoFilter, PairedNoFilter, TooShortReadFilter, TooLongReadFilter,
-	DiscardTrimmedFilter, DiscardUntrimmedFilter, Demultiplexer, NContentFilter)
+	DiscardTrimmedFilter, DiscardUntrimmedFilter, PairedEndDemultiplexer, Demultiplexer,
+	NContentFilter)
 
 
 def safe_divide(numerator, denominator):
@@ -72,7 +73,7 @@ class Statistics:
 
 		# Collect statistics from writers/filters
 		for w in writers:
-			if isinstance(w, (NoFilter, PairedNoFilter, Demultiplexer)) or \
+			if isinstance(w, (NoFilter, PairedNoFilter, PairedEndDemultiplexer, Demultiplexer)) or \
 					isinstance(w.filter, (DiscardTrimmedFilter, DiscardUntrimmedFilter)):
 				self.written += w.written
 				self.written_bp[0] += w.written_bp[0]